+++ util.go
@@ -22,16 +22,16 @@ const (
 
 type Hash [HashLength]byte
 
-func bitIsSet(bits []byte, i uint64) bool {
+func bitIsSet(bits []byte, i int) bool {
 	return bits[i/8]&(1<<uint(7-i%8)) != 0
 }
-func bitSet(bits []byte, i uint64) {
+func bitSet(bits []byte, i int) {
 	bits[i/8] |= 1 << uint(7-i%8)
 }
-func bitUnSet(bits []byte, i uint64) {
+func bitUnSet(bits []byte, i int) {
 	bits[i/8] = bits[i/8] &^ (1 << uint(7-i%8))
 }
-func bitSplit(bits []byte, i uint64) (split []byte) {
+func bitSplit(bits []byte, i int) (split []byte) {
 	split = make([]byte, len(bits))
 	copy(split, bits)
 	bitSet(split, i)
